@@ -1,10 +1,16 @@
 {
   "name": "seefit",
-  "version": "1.1.0",
+  "version": "1.0.0",
+  "author": "Inovretail",
+  "description": "SeeFit Frontend client",
   "scripts": {
     "ng": "ng",
-    "start:dev": "ng serve --host=0.0.0.0 --disable-host-check --proxy-config ./src/proxy.conf.js --poll=2000",
+    "start:dev": "ng serve --host=0.0.0.0 --disable-host-check --proxy-config ./src/proxy.conf.js --poll=2000 ",
     "start:tst": "ng serve --proxy-config ./src/proxy.conf.js --poll=2000",
+    "start:ios": "npx cap open ios",
+    "start:android": "npx cap open android",
+    "start:desktop": "cd electron && npm run electron:start-live",
+    "start:desktop:build": "cd electron && npm run electron:pack",
     "prebuild": "npm run lint:check",
     "build": "ng build -c production",
     "build:native": "rm -rf ../native/www && npm run build && mv dist ../native/www",
@@ -23,33 +29,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.2.0",
-    "@angular/cdk": "^14.2.2",
-    "@angular/common": "^14.2.0",
-    "@angular/compiler": "^14.2.0",
-    "@angular/core": "^14.2.0",
-    "@angular/forms": "^14.2.0",
-    "@angular/material": "^14.2.2",
-    "@angular/platform-browser": "^14.2.0",
-    "@angular/platform-browser-dynamic": "^14.2.0",
-    "@angular/router": "^14.2.0",
-    "@fortawesome/fontawesome-free": "^6.2.0",
-    "@i9r/ui-components": "^1.6.0",
-    "@ngneat/transloco": "^4.1.1",
-    "bootstrap": "^5.2.1",
-    "jquery": "^3.6.1",
+    "@angular/common": "^15.0.0",
+    "@angular/core": "^15.0.0",
+    "@angular/forms": "^15.0.0",
+    "@angular/platform-browser": "^15.0.0",
+    "@angular/platform-browser-dynamic": "^15.0.0",
+    "@angular/router": "^15.0.0",
+    "@capacitor-community/electron": "^4.1.2",
+    "@capacitor/android": "^4.7.1",
+    "@capacitor/app": "^4.1.1",
+    "@capacitor/core": "^4.7.1",
+    "@capacitor/haptics": "^4.1.0",
+    "@capacitor/ios": "^4.7.1",
+    "@capacitor/keyboard": "^4.1.1",
+    "@capacitor/status-bar": "^4.1.1",
+    "@i9r/ui-components": "^2.0.0",
+    "@ionic/angular": "^7.0.0",
+    "@ngx-translate/core": "^14.0.0",
+    "@ngx-translate/http-loader": "^7.0.0",
+    "ionicons": "^7.1.0",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.2.2",
-    "@angular/cli": "~14.2.2",
-    "@angular/compiler-cli": "^14.2.0",
-    "@ngneat/transloco-keys-manager": "^3.4.2",
-    "@types/node": "^18.7.18",
+    "@angular-devkit/build-angular": "^15.0.0",
+    "@angular-eslint/builder": "^15.0.0",
+    "@angular-eslint/eslint-plugin": "^15.0.0",
+    "@angular-eslint/eslint-plugin-template": "^15.0.0",
+    "@angular-eslint/schematics": "^15.0.0",
+    "@angular-eslint/template-parser": "^15.0.0",
+    "@angular/cli": "^15.0.0",
+    "@angular/compiler": "^15.0.0",
+    "@angular/compiler-cli": "^15.0.0",
+    "@angular/language-service": "^15.0.0",
+    "@capacitor/cli": "^4.7.1",
+    "@ionic/angular-toolkit": "^8.0.0",
+    "@types/jasmine": "~4.0.0",
+    "@types/node": "^12.11.1",
+    "@typescript-eslint/eslint-plugin": "5.3.0",
+    "@typescript-eslint/parser": "5.3.0",
+    "eslint": "^7.6.0",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsdoc": "30.7.6",
+    "eslint-plugin-prefer-arrow": "1.2.2",
     "cypress": "^10.8.0",
     "prettier": "^2.8.4",
-    "typescript": "~4.7.2"
+    "ts-node": "~8.3.0",
+    "typescript": "~4.8.4"
   }
 }
