@@ -1,16 +1,15 @@
 import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { AuthGuardGuard } from './core/guards/auth-guard.guard';
-import { UnAuthGuardGuard } from './core/guards/un-auth-guard.guard';
+import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
+import { AuthGuard } from './core/guards/auth.guard';
 
 const routes: Routes = [
   {
     path: '',
-    redirectTo: 'authentication',
+    redirectTo: 'auth',
     pathMatch: 'full',
   },
   {
-    path: 'authentication',
+    path: 'auth',
     loadChildren: () =>
       import('./modules/authentication/authentication.module').then(
         m => m.AuthenticationModule
@@ -18,14 +17,20 @@ const routes: Routes = [
   },
   {
     path: 'platform',
-    canActivateChild: [AuthGuardGuard],
+    canActivate: [AuthGuard],
     loadChildren: () =>
       import('./modules/platform/platform.module').then(m => m.PlatformModule),
+  },
+  {
+    path: '**',
+    redirectTo: '/platform',
   },
 ];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [
+    RouterModule.forRoot(routes, { preloadingStrategy: PreloadAllModules }),
+  ],
   exports: [RouterModule],
 })
 export class AppRoutingModule {}
