@@ -118,13 +118,13 @@ export class DataService {
   getDummyData(): Array<ICIData> {
     return [
       {
-        title: 'Abby Cannon',
-        userId: 'N#7187',
+        title: 'Sarah Holloway',
+        userId: 'N#3929',
         status: INTERACTION_STATUS.PLANNED,
         inClub: true,
         excludeAG: false,
 
-        date: '2023-04-13',
+        date: '2023-03-20',
         interaction: {
           label: 'INTERACTION:',
           value: 'First Day',
@@ -141,8 +141,9 @@ export class DataService {
           value: 'n/a',
           color: '#000000',
         },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_female3.jpg',
-        interactionInfo: '',
+        imageUrl: 'assets/mockData/images/userPhotos/profile_female1.jpg',
+        interactionInfo:
+          'Sarah Holloway grade our service 5 (1-10 scale), which puts her as a detractor. Related to gym floor satisfaction, the evaluation is 8.',
         customerInfo: {
           customerRecord: {
             title: 'interaction.data.customer-info',
@@ -691,11 +692,11 @@ export class DataService {
       },
 
       {
-        title: 'Elle Lawrence',
-        userId: 'N#4929',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: false,
-        excludeAG: true,
+        title: 'Alan Rivers',
+        userId: 'N#3203',
+        status: INTERACTION_STATUS.COMPLETED,
+        inClub: true,
+        excludeAG: false,
 
         date: '2023-04-13',
         interaction: {
@@ -714,7 +715,7 @@ export class DataService {
           value: 'n/a',
           color: '#000000',
         },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_male4.jpg',
+        imageUrl: 'assets/mockData/images/userPhotos/profile_male2.jpg',
         interactionInfo:
           "Edmund Jacobson started his membership on 2023-03-09, but only visit one time in the first 2 weeks. It's important to sensibilize the customer regarding the importance of the onboarding, namely by the definition of a weekly schedule of workouts (at least twice a week)",
         customerInfo: {
@@ -982,8 +983,8 @@ export class DataService {
       },
 
       {
-        title: 'Josh Wolf',
-        userId: 'N#3701',
+        title: 'Jana Miller',
+        userId: 'N#3204',
         status: INTERACTION_STATUS.COMPLETED,
         inClub: true,
         excludeAG: false,
@@ -1005,8 +1006,9 @@ export class DataService {
           value: 'Asleep',
           color: '#000000',
         },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_male7.jpg',
-        interactionInfo: '',
+        imageUrl: 'assets/mockData/images/userPhotos/profile_female2.jpg',
+        interactionInfo:
+          'Jana Miller had an increase of risk in 8pp in the last 14 days (currently at 26%). As this is generated by an algorithm, check the customer tab for deep-diving purposes',
         customerInfo: {
           customerRecord: {
             title: 'interaction.data.customer-info',
@@ -1407,9 +1409,9 @@ export class DataService {
       },
 
       {
-        title: 'Jana Miller',
-        userId: 'N#3204',
-        status: INTERACTION_STATUS.PLANNED,
+        title: 'Rupert Horton',
+        userId: 'N#2390',
+        status: INTERACTION_STATUS.COMPLETED,
         inClub: true,
         excludeAG: false,
 
@@ -1430,9 +1432,8 @@ export class DataService {
           value: 'Moderate',
           color: '#000000',
         },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_female2.jpg',
-        interactionInfo:
-          'Jana Miller had an increase of risk in 8pp in the last 14 days (currently at 26%). As this is generated by an algorithm, check the customer tab for deep-diving purposes',
+        imageUrl: 'assets/mockData/images/userPhotos/profile_male3.jpg',
+        interactionInfo: '',
         customerInfo: {
           customerRecord: {
             title: 'interaction.data.customer-info',
@@ -1836,10 +1837,163 @@ export class DataService {
       },
 
       {
-        title: 'Phillip Ponce',
-        userId: 'N#674',
-        status: INTERACTION_STATUS.COMPLETED,
+        title: 'Abby Cannon',
+        userId: 'N#7187',
+        status: INTERACTION_STATUS.PLANNED,
         inClub: true,
+        excludeAG: true,
+
+        date: '2023-03-21',
+        interaction: {
+          label: 'INTERACTION:',
+          value: 'First Day',
+          color: '#000000',
+          isBold: true,
+        },
+        primary: {
+          label: 'RISK:',
+          value: 'n/a',
+          color: '#000000',
+        },
+        secondary: {
+          label: 'USAGE:',
+          value: 'n/a',
+          color: '#000000',
+        },
+        imageUrl: 'assets/mockData/images/userPhotos/profile_female3.jpg',
+        interactionInfo: '',
+        customerInfo: {
+          customerRecord: {
+            title: 'interaction.data.customer-info',
+            content: [
+              {
+                key: 'interaction.data.birth-date',
+                value: '2000-04-22 (22)',
+              },
+              {
+                key: 'interaction.data.objective',
+                value: 'Gain weight/muscle',
+              },
+              {
+                key: 'interaction.data.access',
+                value: '0/0/0/0',
+              },
+              {
+                key: 'interaction.data.ag',
+                value: '0/0/0/0',
+              },
+              {
+                key: 'interaction.data.shareAG',
+                value: '0%',
+              },
+              {
+                key: 'interaction.data.last-af',
+                value: 'n/a',
+              },
+              {
+                key: 'interaction.data.visit-length',
+                value: 'n/a',
+              },
+            ],
+          },
+          contractType: {
+            title: 'interaction.data.contract-info',
+            content: [
+              {
+                key: 'interaction.data.membership',
+                value: '2023-03-17 (1)',
+              },
+              {
+                key: 'interaction.data.subscription-type',
+                value: '9.9$ / Week',
+              },
+              {
+                key: 'interaction.data.schedule',
+                value: 'Unrestricted',
+              },
+              {
+                key: 'interaction.data.additional-services',
+                value: 'n/a',
+              },
+            ],
+          },
+          additionalInfo: {
+            title: 'interaction.data.additional-info',
+            icons: [
+              'ic-1_1',
+              'ic-1_2',
+              'ic-1_3',
+              'ic-1_4',
+              'ic-1_5',
+              'ic-2_1',
+              'ic-3_1',
+              'ic-3_2',
+              'ic-3_3',
+              'ic-3_4',
+              'ic-4_1',
+              'ic-4_2',
+              'ic-4_3',
+              'ic-4_4',
+              'ic-4_5',
+            ],
+            lifeCycle: {
+              key: 'interaction.data.lifecycle',
+              progress: 0,
+              topLabels: [
+                {
+                  label: '0',
+                  position: 0,
+                },
+                {
+                  label: '45',
+                  position: 25,
+                },
+                {
+                  label: '90',
+                  position: 50,
+                },
+                {
+                  label: '135',
+                  position: 75,
+                },
+                {
+                  label: '180',
+                  position: 95,
+                },
+              ],
+              bottomLabels: [
+                {
+                  label: 'ONB1',
+                  position: 0,
+                },
+                {
+                  label: 'REP1',
+                  position: 25,
+                },
+                {
+                  label: 'REP2',
+                  position: 50,
+                },
+                {
+                  label: 'AUT1',
+                  position: 75,
+                },
+                {
+                  label: 'AUT2',
+                  position: 95,
+                },
+              ],
+            },
+          },
+          observation: '',
+        },
+      },
+
+      {
+        title: 'Alice Williamson',
+        userId: 'N#4812',
+        status: INTERACTION_STATUS.COMPLETED,
+        inClub: false,
         excludeAG: false,
 
         date: '2023-03-15',
@@ -1989,10 +2143,10 @@ export class DataService {
       },
 
       {
-        title: 'Sarah Holloway',
-        userId: 'N#3929',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: true,
+        title: 'Tim Shepard',
+        userId: 'N#9027',
+        status: INTERACTION_STATUS.COMPLETED,
+        inClub: false,
         excludeAG: false,
         date: '2023-04-13',
         interaction: {
@@ -2134,10 +2288,10 @@ export class DataService {
       },
 
       {
-        title: 'Marion Brewer',
-        userId: 'N#914',
+        title: 'Helena Saunders',
+        userId: 'N#5653',
         status: INTERACTION_STATUS.COMPLETED,
-        inClub: true,
+        inClub: false,
         excludeAG: false,
         date: '2023-04-05',
         interaction: {
@@ -2704,581 +2858,9 @@ export class DataService {
       },
 
       {
-        title: 'Sanaa Tyler',
-        userId: 'N#5653',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: true,
-        excludeAG: true,
-
-        date: '2023-02-28',
-        interaction: {
-          label: 'INTERACTION:',
-          value: 'NPS Promoter',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '9%',
-          color: '#00FF00',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Active',
-          color: '#000000',
-        },
-        imageUrl: 'assets/icons/no-avatar.svg',
-        interactionInfo: '',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1981-12-08 (41)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Improve well-being',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '0/9/19/118',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/9/19/116',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '99%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2023-01-29 (51)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '46 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2022-03-08 (12)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '29$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: '07:00 - 18:00 | 20:30 - 22:00',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'n/a',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: [
-              'ic-1_1',
-              'ic-1_2',
-              'ic-1_3',
-              'ic-1_4',
-              'ic-1_5',
-              'ic-2_1',
-              'ic-3_1',
-              'ic-3_2',
-              'ic-3_3',
-              'ic-3_4',
-              'ic-4_1',
-              'ic-4_2',
-              'ic-4_3',
-              'ic-4_4',
-              'ic-4_5',
-            ],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Alan Rivers',
-        userId: 'N#3203',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: true,
-        excludeAG: true,
-
-        date: '2023-04-13',
-        interaction: {
-          label: 'INTERACTION:',
-          value: 'Cycle Completed',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '5%',
-          color: '#00FF00',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Very Active',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_male2.jpg',
-        interactionInfo:
-          "Alan Rivers has completed the REP_2 program with 13 visits on 4 weeks. It's time to congratulate the member, pushing for even higher usage in the following weeks.",
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1966-10-22 (68)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Improve well-being',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '3/13/25/137',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '1/5/11/46',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '43%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2023-01-17 (63)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '62 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2021-01-30 (26)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '9.9$ / Week',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: 'Unrestricted',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: '1 PT session / week',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: ['ic-1_5', 'ic-2_1', 'ic-3_1', 'ic-3_2', 'ic-3_3'],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Alan Rivers',
-        userId: 'N#3203',
+        title: 'Walter Wiggins',
+        userId: 'N#1903',
         status: INTERACTION_STATUS.COMPLETED,
-        inClub: true,
-        excludeAG: false,
-
-        date: '2023-04-05',
-        interaction: {
-          label: 'LAST INTERACTION:',
-          value: 'High Motivation',
-          color: '#000000',
-          isBold: !true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '7%',
-          color: '#00FF00',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Very Active',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_male2.jpg',
-        interactionInfo:
-          "Alan Rivers has completed the REP_2 program with 13 visits on 4 weeks. It's time to congratulate the member, pushing for even higher usage in the following weeks.",
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1966-10-22 (68)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Improve well-being',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '3/13/25/137',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '1/5/11/46',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '43%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2023-01-17 (63)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '62 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2021-01-30 (26)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '9.9$ / Week',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: 'Unrestricted',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: '1 PT session / week',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: ['ic-1_5', 'ic-2_1', 'ic-3_1', 'ic-3_2', 'ic-3_3'],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-      {
-        title: 'Abby Cannon',
-        userId: 'N#234234',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: false,
-        excludeAG: false,
-        date: '2023-04-13',
-        interaction: {
-          label: 'INTERACTION:',
-          value: 'First Day',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: 'n/a',
-          color: '#000000',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'n/a',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_female3.jpg',
-        interactionInfo: '',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '2000-04-22 (22)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Gain weight/muscle',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '0/0/0/0',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/0/0/0',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '0%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: 'n/a',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: 'n/a',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2023-03-17 (1)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '9.9$ / Week',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: 'Unrestricted',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'n/a',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: [
-              'ic-1_1',
-              'ic-1_2',
-              'ic-1_5',
-              'ic-2_1',
-              'ic-3_1',
-              'ic-4_4',
-              'ic-4_5',
-            ],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 0,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Laura Ashley',
-        userId: 'N#3422',
-        status: INTERACTION_STATUS.PLANNED,
         inClub: false,
         excludeAG: false,
 
@@ -4246,1883 +3828,6 @@ export class DataService {
               'ic-4_4',
               'ic-4_5',
             ],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Walter Wiggins',
-        userId: 'N#82323',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: false,
-        excludeAG: false,
-        date: '2023-04-13',
-        interaction: {
-          label: 'INTERACTION:',
-          value: 'Medium Churn Risk',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '28%',
-          color: '#FFA500',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Low',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_male8.jpg',
-        interactionInfo: '',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1960-04-22 (62)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Improve well-being',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '0/0/2/7',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/0/0/0',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '0%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: 'n/a',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '69 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2021-01-30 (26)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '24$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: '07:00 - 17:00',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'n/a',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: [
-              'ic-1_1',
-              'ic-1_2',
-              'ic-1_3',
-              'ic-1_4',
-              'ic-1_5',
-              'ic-2_1',
-              'ic-3_1',
-              'ic-3_2',
-              'ic-3_3',
-              'ic-3_4',
-              'ic-4_1',
-              'ic-4_2',
-              'ic-4_3',
-              'ic-4_4',
-              'ic-4_5',
-            ],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Jana Miller',
-        userId: 'N#82323',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: false,
-        excludeAG: false,
-
-        date: '2023-04-13',
-        interaction: {
-          label: 'INTERACTION:',
-          value: 'Churn Risk Increase',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '26%',
-          color: '#FFA500',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Moderate',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_female2.jpg',
-        interactionInfo:
-          'Jana Miller had an increase of risk in 8pp in the last 14 days (currently at 26%). As this is generated by an algorithm, check the customer tab for deep-diving purposes',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1990-08-05 (32)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Gain weight/muscle',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '0/4/10/88',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/0/0/13',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '15%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: 'n/a',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '53 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2021-03-30 (24)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '24$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: '07:00 - 17:00',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'n/a',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: ['ic-1_4', 'ic-1_5', 'ic-2_1', 'ic-3_1'],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Norma Rollins',
-        userId: 'N#744423',
-        status: INTERACTION_STATUS.COMPLETED,
-        inClub: false,
-        excludeAG: false,
-
-        date: '2023-04-07',
-        interaction: {
-          label: 'LAST INTERACTION:',
-          value: 'SMS - New Group Class',
-          color: '#000000',
-          isBold: !true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '25%',
-          color: '#FFA500',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Moderate',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_female2.jpg',
-        interactionInfo:
-          'Jana Miller had an increase of risk in 8pp in the last 14 days (currently at 26%). As this is generated by an algorithm, check the customer tab for deep-diving purposes',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1990-08-05 (32)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Gain weight/muscle',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '0/4/10/88',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/0/0/13',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '15%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: 'n/a',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '53 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2021-03-30 (24)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '24$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: '07:00 - 17:00',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'n/a',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: ['ic-1_4', 'ic-1_5', 'ic-2_1', 'ic-3_1'],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Alice Williamson',
-        userId: 'N#73333',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: false,
-        excludeAG: false,
-
-        date: '2023-04-13',
-        interaction: {
-          label: 'INTERACTION:',
-          value: 'Low Motivation',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '21%',
-          color: '#FFA500',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Moderate',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_female8.jpg',
-        interactionInfo: '',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1971-11-02 (51)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Weight Loss',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '1/5/9/85',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/0/0/1',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '1%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2022-08-09 (201)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '82 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2018-09-23 (54)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '22.5$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: 'Unrestricted',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'Nutrition Plan',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: [
-              'ic-1_1',
-              'ic-1_2',
-              'ic-1_3',
-              'ic-1_4',
-              'ic-1_5',
-              'ic-2_1',
-              'ic-3_1',
-              'ic-3_2',
-              'ic-3_3',
-              'ic-3_4',
-              'ic-4_1',
-              'ic-4_2',
-              'ic-4_3',
-              'ic-4_4',
-              'ic-4_5',
-            ],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-      {
-        title: 'Phillip Ponce',
-        userId: 'N#63333',
-        status: INTERACTION_STATUS.COMPLETED,
-        inClub: false,
-        excludeAG: false,
-
-        date: '2023-03-15',
-        interaction: {
-          label: 'LAST INTERACTION:',
-          value: 'Planned Next Visit',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '21%',
-          color: '#FFA500',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Moderate',
-          color: '#000000',
-        },
-        imageUrl: 'assets/icons/no-avatar.svg',
-        interactionInfo: '',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1971-11-02 (51)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Weight Loss',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '1/5/9/85',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/0/0/1',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '1%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2022-08-09 (201)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '82 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2018-09-23 (54)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '22.5$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: 'Unrestricted',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'Nutrition Plan',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: [
-              'ic-1_1',
-              'ic-1_2',
-              'ic-1_3',
-              'ic-1_4',
-              'ic-1_5',
-              'ic-2_1',
-              'ic-3_1',
-              'ic-3_2',
-              'ic-3_3',
-              'ic-3_4',
-              'ic-4_1',
-              'ic-4_2',
-              'ic-4_3',
-              'ic-4_4',
-              'ic-4_5',
-            ],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Sarah Holloway',
-        userId: 'N#7334',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: false,
-        excludeAG: false,
-        date: '2023-04-13',
-        interaction: {
-          label: 'INTERACTION:',
-          value: 'NPS Detractor',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '19%',
-          color: '#cccc00',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Active',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_female1.jpg',
-        interactionInfo:
-          'Sarah Holloway grade our service 5 (1-10 scale), which puts her as a detractor. Related to gym floor satisfaction, the evaluation is 8.',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1982-02-08 (41)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Weight Loss',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '2/7/19/97',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/0/1/3',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '3%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2022-12-14 (97)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '57 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2020-11-30 (28)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '29$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: '07:00 - 18:00 | 20:30 - 22:00',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'n/a',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: [
-              'ic-1_4',
-              'ic-1_5',
-              'ic-2_1',
-              'ic-3_1',
-              'ic-3_2',
-              'ic-4_4',
-              'ic-4_5',
-            ],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Marion Brewer',
-        userId: 'N#3334',
-        status: INTERACTION_STATUS.COMPLETED,
-        inClub: false,
-        excludeAG: false,
-        date: '2023-04-05',
-        interaction: {
-          label: 'LAST INTERACTION:',
-          value: 'Monthly Challenge',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '17%',
-          color: '#cccc00',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Moderate',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_female1.jpg',
-        interactionInfo:
-          'Sarah Holloway grade our service 5 (1-10 scale), which puts her as a detractor. Related to gym floor satisfaction, the evaluation is 8.',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1982-02-08 (41)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Weight Loss',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '2/7/19/97',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/0/1/3',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '3%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2022-12-14 (97)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '57 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2020-11-30 (28)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '29$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: '07:00 - 18:00 | 20:30 - 22:00',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'n/a',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: [
-              'ic-1_4',
-              'ic-1_5',
-              'ic-2_1',
-              'ic-3_1',
-              'ic-3_2',
-              'ic-4_4',
-              'ic-4_5',
-            ],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Rupert Horton',
-        userId: 'N#76556',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: false,
-        excludeAG: false,
-        date: '2023-04-13',
-        interaction: {
-          label: 'INTERACTION:',
-          value: 'Usage Very Active -> Active',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '11%',
-          color: '#000000',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Active',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_male3.jpg',
-        interactionInfo: '',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1982-06-29 (40)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Improve well-being',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '1/10/24/203',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/0/0/0',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '0%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2022-10-03 (169)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '58 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2021-07-08 (20)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '39$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: 'Unrestricted',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'n/a',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: ['ic-1_2', 'ic-1_3', 'ic-1_4', 'ic-1_5', 'ic-4_5'],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Rebecca Mcleod',
-        userId: 'N#4234',
-        status: INTERACTION_STATUS.COMPLETED,
-        inClub: false,
-        excludeAG: false,
-        date: '2023-04-08',
-        interaction: {
-          label: 'LAST INTERACTION:',
-          value: 'Email - 100 visits ',
-          color: '#000000',
-          isBold: !true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '11%',
-          color: '#000000',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Active',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_male3.jpg',
-        interactionInfo: '',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1982-06-29 (40)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Improve well-being',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '1/10/24/203',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/0/0/0',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '0%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2022-10-03 (169)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '58 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2021-07-08 (20)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '39$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: 'Unrestricted',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'n/a',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: ['ic-1_2', 'ic-1_3', 'ic-1_4', 'ic-1_5', 'ic-4_5'],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Helena Saunders',
-        userId: 'N#5676767',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: false,
-        excludeAG: false,
-
-        date: '2023-04-13',
-        interaction: {
-          label: 'INTERACTION:',
-          value: 'Cycle 75% Under',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '9%',
-          color: '#00FF00',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Active',
-          color: '#000000',
-        },
-        imageUrl: 'assets/icons/no-avatar.svg',
-        interactionInfo: '',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1981-12-08 (41)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Improve well-being',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '0/9/19/118',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/9/19/116',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '99%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2023-01-29 (51)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '46 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2022-03-08 (12)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '29$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: '07:00 - 18:00 | 20:30 - 22:00',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'n/a',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: [
-              'ic-1_1',
-              'ic-1_2',
-              'ic-1_3',
-              'ic-1_4',
-              'ic-1_5',
-              'ic-2_1',
-              'ic-3_1',
-              'ic-3_2',
-              'ic-3_3',
-              'ic-3_4',
-              'ic-4_1',
-              'ic-4_2',
-              'ic-4_3',
-              'ic-4_4',
-              'ic-4_5',
-            ],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Sanaa Tyler',
-        userId: 'N#767667',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: false,
-        excludeAG: false,
-
-        date: '2023-02-28',
-        interaction: {
-          label: 'INTERACTION:',
-          value: 'NPS Promoter',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '9%',
-          color: '#00FF00',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Active',
-          color: '#000000',
-        },
-        imageUrl: 'assets/icons/no-avatar.svg',
-        interactionInfo: '',
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1981-12-08 (41)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Improve well-being',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '0/9/19/118',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '0/9/19/116',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '99%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2023-01-29 (51)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '46 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2022-03-08 (12)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '29$ / Month',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: '07:00 - 18:00 | 20:30 - 22:00',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: 'n/a',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: [
-              'ic-1_1',
-              'ic-1_2',
-              'ic-1_3',
-              'ic-1_4',
-              'ic-1_5',
-              'ic-2_1',
-              'ic-3_1',
-              'ic-3_2',
-              'ic-3_3',
-              'ic-3_4',
-              'ic-4_1',
-              'ic-4_2',
-              'ic-4_3',
-              'ic-4_4',
-              'ic-4_5',
-            ],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Alan Rivers',
-        userId: 'N#377767',
-        status: INTERACTION_STATUS.PLANNED,
-        inClub: false,
-        excludeAG: false,
-
-        date: '2023-04-13',
-        interaction: {
-          label: 'INTERACTION:',
-          value: 'Cycle Completed',
-          color: '#000000',
-          isBold: true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '5%',
-          color: '#00FF00',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Very Active',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_male2.jpg',
-        interactionInfo:
-          "Alan Rivers has completed the REP_2 program with 13 visits on 4 weeks. It's time to congratulate the member, pushing for even higher usage in the following weeks.",
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1966-10-22 (68)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Improve well-being',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '3/13/25/137',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '1/5/11/46',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '43%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2023-01-17 (63)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '62 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2021-01-30 (26)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '9.9$ / Week',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: 'Unrestricted',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: '1 PT session / week',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: ['ic-1_5', 'ic-2_1', 'ic-3_1', 'ic-3_2', 'ic-3_3'],
-            lifeCycle: {
-              key: 'interaction.data.lifecycle',
-              progress: 75,
-              topLabels: [
-                {
-                  label: '0',
-                  position: 0,
-                },
-                {
-                  label: '45',
-                  position: 25,
-                },
-                {
-                  label: '90',
-                  position: 50,
-                },
-                {
-                  label: '135',
-                  position: 75,
-                },
-                {
-                  label: '180',
-                  position: 95,
-                },
-              ],
-              bottomLabels: [
-                {
-                  label: 'ONB1',
-                  position: 0,
-                },
-                {
-                  label: 'REP1',
-                  position: 25,
-                },
-                {
-                  label: 'REP2',
-                  position: 50,
-                },
-                {
-                  label: 'AUT1',
-                  position: 75,
-                },
-                {
-                  label: 'AUT2',
-                  position: 95,
-                },
-              ],
-            },
-          },
-          observation: '',
-        },
-      },
-
-      {
-        title: 'Alan Rivers',
-        userId: 'N#98080',
-        status: INTERACTION_STATUS.COMPLETED,
-        inClub: false,
-        excludeAG: false,
-
-        date: '2023-04-05',
-        interaction: {
-          label: 'LAST INTERACTION:',
-          value: 'High Motivation',
-          color: '#000000',
-          isBold: !true,
-        },
-        primary: {
-          label: 'RISK:',
-          value: '7%',
-          color: '#00FF00',
-        },
-        secondary: {
-          label: 'USAGE:',
-          value: 'Very Active',
-          color: '#000000',
-        },
-        imageUrl: 'assets/mockData/images/userPhotos/profile_male2.jpg',
-        interactionInfo:
-          "Alan Rivers has completed the REP_2 program with 13 visits on 4 weeks. It's time to congratulate the member, pushing for even higher usage in the following weeks.",
-        customerInfo: {
-          customerRecord: {
-            title: 'interaction.data.customer-info',
-            content: [
-              {
-                key: 'interaction.data.birth-date',
-                value: '1966-10-22 (68)',
-              },
-              {
-                key: 'interaction.data.objective',
-                value: 'Improve well-being',
-              },
-              {
-                key: 'interaction.data.access',
-                value: '3/13/25/137',
-              },
-              {
-                key: 'interaction.data.ag',
-                value: '1/5/11/46',
-              },
-              {
-                key: 'interaction.data.shareAG',
-                value: '43%',
-              },
-              {
-                key: 'interaction.data.last-af',
-                value: '2023-01-17 (63)',
-              },
-              {
-                key: 'interaction.data.visit-length',
-                value: '62 min',
-              },
-            ],
-          },
-          contractType: {
-            title: 'interaction.data.contract-info',
-            content: [
-              {
-                key: 'interaction.data.membership',
-                value: '2021-01-30 (26)',
-              },
-              {
-                key: 'interaction.data.subscription-type',
-                value: '9.9$ / Week',
-              },
-              {
-                key: 'interaction.data.schedule',
-                value: 'Unrestricted',
-              },
-              {
-                key: 'interaction.data.additional-services',
-                value: '1 PT session / week',
-              },
-            ],
-          },
-          additionalInfo: {
-            title: 'interaction.data.additional-info',
-            icons: ['ic-1_5', 'ic-2_1', 'ic-3_1', 'ic-3_2', 'ic-3_3'],
             lifeCycle: {
               key: 'interaction.data.lifecycle',
               progress: 75,
