@@ -1,7 +1,4 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
-import { MatSidenav } from '@angular/material/sidenav';
-import { ActivatedRoute, Route, Router } from '@angular/router';
-import { AuthService, UserEntity } from 'src/app/core/services/auth.service';
+import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-platform',
@@ -9,155 +6,17 @@ import { AuthService, UserEntity } from 'src/app/core/services/auth.service';
   styleUrls: ['./platform.component.scss'],
 })
 export class PlatformComponent implements OnInit {
-  @ViewChild('sidenav') sidenav: MatSidenav | undefined;
-
-  permissions = [
-    // object structure should represent the actual menu (menu/page)
-    {
-      appCode: 'Interactions', // this block has no connection to the app structure
-      icon: 'interaction.svg',
-      desc: [
-        {
-          lang: 'pt',
-          text: 'Interações',
-        },
-        {
-          lang: 'en',
-          text: 'Interactions',
-        },
-      ],
-      permissions: [
-        {
-          // permissions: [A, B, C] - check if user has permission
-          // app: X - check if origin has access to app
-          permissionCode: 'ICI',
-          desc: [
-            {
-              lang: 'pt',
-              text: 'Interações em Ginásio',
-            },
-            {
-              lang: 'en',
-              text: 'In club interactions',
-            },
-          ],
-        },
-        {
-          permissionCode: 'OCI',
-          desc: [
-            {
-              lang: 'pt',
-              text: 'Interações fora Ginásio',
-            },
-            {
-              lang: 'en',
-              text: 'Off club interactions',
-            },
-          ],
-        },
-        {
-          permissionCode: 'POWER_BI',
-        },
-      ],
-    },
-    {
-      appCode: 'Trainning',
-      icon: 'training.svg',
-      desc: [
-        {
-          lang: 'pt',
-          text: 'Plano de treino',
-        },
-        {
-          lang: 'en',
-          text: 'Training Plan',
-        },
-      ],
-      permissions: [
-        {
-          permissionCode: 'AF',
-          desc: [
-            {
-              lang: 'pt',
-              text: 'Interações em Ginásio',
-            },
-            {
-              lang: 'en',
-              text: 'In club interactions',
-            },
-          ],
-        },
-        {
-          permissionCode: 'PT',
-          desc: [
-            {
-              lang: 'pt',
-              text: 'Interações fora Ginásio',
-            },
-            {
-              lang: 'en',
-              text: 'Off club interactions',
-            },
-          ],
-        },
-        {
-          permissionCode: 'POWER_BI',
-        },
-      ],
-    },
+  public appPages = [
+    //TODO: refactor menu options get from service
+    { title: 'Inbox', url: '/folder/Inbox', icon: 'mail' },
+    { title: 'Outbox', url: '/folder/Outbox', icon: 'paper-plane' },
+    { title: 'Favorites', url: '/folder/Favorites', icon: 'heart' },
+    { title: 'Archived', url: '/folder/Archived', icon: 'archive' },
+    { title: 'Trash', url: '/folder/Trash', icon: 'trash' },
+    { title: 'Spam', url: '/folder/Spam', icon: 'warning' },
   ];
-  opened = false;
-  childActivated = false;
-  originName = 'Solinca';
+  public labels = ['Family', 'Friends', 'Notes', 'Work', 'Travel', 'Reminders'];
+  constructor() {}
 
-  user: UserEntity | undefined;
-
-  constructor(
-    private router: Router,
-    private activatedRoute: ActivatedRoute,
-    private auth: AuthService
-  ) {}
-
-  ngOnInit(): void {
-    this.user = this.auth.getUserInfo();
-  }
-
-  sideNavVisibily() {
-    this.router.navigate(['/']);
-    return;
-    this.sidenav?.toggle();
-  }
-
-  setSidenavStatus(event: boolean) {
-    this.opened = event;
-  }
-
-  /**
-   * on username dropdown click
-   */
-  usernameClick() {
-    this.router.navigate([`./user_profile/${this.user?.userId}`], {
-      relativeTo: this.activatedRoute,
-    });
-  }
-
-  /**
-   * on origin dropdown click
-   */
-  originClick() {}
-
-  /**
-   * on route child activated
-   */
-  onActivateChild() {
-    this.setSidenavStatus(false);
-    this.childActivated = true;
-  }
-
-  /**
-   * on route child deactivated
-   */
-  onDeactivateChild() {
-    this.childActivated = false;
-  }
+  ngOnInit(): void {}
 }
