@@ -1,4 +1,5 @@
 export const environment = {
-  endpoint: '/middleware',
+  baseUrl: '/middleware',
+  i18nURL: `http://127.0.0.1:8080`, // `../assets/i18n`
   production: false,
 };
